clientside/requirements.txt
python-dotenv==0.21.0
streamlit==1.12.0
watchdog==2.1.9
pydantic==1.10.2
pandas = ==1.3.3
python-docx==0.8.11
xfilios==0.0.4